[Mono.Android] Support legacy RegisterNativeMembers in trimmable typemap
Add an opt-in feature switch,
`Microsoft.Android.Runtime.RuntimeFeature.LegacyJniRegistration`
(MSBuild `_AndroidEnableLegacyJniRegistration`, default `false`), that
re-enables the legacy, reflection-based JNI native-method registration in
the trimmable type map path.

By default the trimmable type map registers native methods via the fast
path (JCW static initializers -> `mono.android.Runtime.registerNatives`),
and `TrimmableTypeMapTypeManager.RegisterNativeMembers` throws. That makes
two legacy mechanisms unusable with the trimmable type map:

* legacy precompiled Java Callable Wrappers (from binding jars/aars) whose
  static initializers call `mono.android.Runtime.register("Type, Asm",
  Class, "methods")`, and
* `Java.Interop.ManagedPeer.registerNativeMembers`.

When the switch is enabled:

* `JNIEnvInit.Initialize` wires `registerJniNativesFn` even in the
  trimmable path, so the native `register(...)` call flows back into
  managed code; and
* `TrimmableTypeMapTypeManager.RegisterNativeMembers` performs the
  reflection-based registration instead of throwing.

The reflection parsing logic is extracted from `ManagedTypeManager` into a
shared `NativeMethodRegistrar` helper and reused by both managers (removing
a duplicate copy). With the switch off (the default), the helper and the
`register(...)` callback are unreferenced and trimmed away.

No native changes are required: `Java_mono_android_Runtime_register` is a
name-exported JNICALL that is always present and only acts when the managed
`registerJniNativesFn` pointer is set.
