+++ src/Mono.Android/Mono.Android.csproj
@@ -365,6 +365,7 @@
     <Compile Include="Microsoft.Android.Runtime\JniRemappingLookup.cs" />
     <Compile Include="Microsoft.Android.Runtime\ManagedTypeManager.cs" />
     <Compile Include="Microsoft.Android.Runtime\ManagedTypeMapping.cs" />
+    <Compile Include="Microsoft.Android.Runtime\NativeMethodRegistrar.cs" />
     <Compile Include="Microsoft.Android.Runtime\JavaMarshalValueManager.cs" />
     <Compile Include="Microsoft.Android.Runtime\RuntimeFeature.cs" />
     <Compile Include="Microsoft.Android.Runtime\SimpleValueManager.cs" />
